requirements.txt
google-cloud-storage==1.44.0
hydra-core==1.1.1
kornia==0.6.2
matplotlib==3.5.0
numpy==1.21.2
omegaconf==2.1.1
Pillow==9.0.0
protobuf==3.19.3
python-dotenv==0.19.2
requests==2.26.0
scikit_learn==1.0.2
setuptools==58.0.4
torch==1.10.1
torchvision==0.11.2
tqdm==4.62.3
wandb==0.12.9
